fix (functionality): modify Data display
Able to identify Value clearly while printing data details.
Able to clearly change data type except for Address type.

Modify and add "Value" in display data details.
It's not expected to mark the data type Empty for all cases.
Change the type to Empty only if the previous type is not
Address-type and storing memory address is non empty.

Resolves: #81, #82